// lib/types.ts
export type MediaItem = {
  poster_path: string;
  adult: boolean;
  overview: string;
  release_date: string;
  genres: string[];
  id: number;
  original_title: string;
  original_language: string;
  title: string;
  backdrop_path: string;
  popularity: number;
  vote_count: number;
  vote: boolean;
  vote_average: number;
};